Move from @deprecated PegDown to Flexmark
This step is forced by the fact, that the last release of PegDown 1.6.0
(3 years old)
uses ASM 5.0.x, that isn't capable of dealing with Java 8 byte code.
Since Xtext 2.14 was compiled with Java8 (or later), we are now forced
to upgrade.

According to the suggestion on the GitHub PegDown page it is recommented
to move to Flexmark.

TODO:
- Write more MarkDown tests
- Verify generated html code
- Check, that code highlighting is working